vector-twin/pyproject.toml
[tool.poetry]
name = "vector-twin"
version = "0.1.0"
description = "Find your Twin Celebrity in Vector Space"
authors = ["MichaelisTrofficus <miguel.otero.pedrido.1993@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
facenet-pytorch = "^2.6.0"
python-dotenv = "^1.0.1"
qdrant-client = "^1.12.1"
datasets = "^3.1.0"
zenml = {extras = ["server"], version = "^0.68.1"}
matplotlib = "^3.9.2"
tqdm = "^4.67.0"
poethepoet = "^0.30.0"
pre-commit = "^4.0.1"
click = "8.1.3"
google-cloud-storage = "^2.18.2"
fastapi-cache = "^0.1.0"
fastapi-cache2 = "^0.2.2"
redis = "^5.2.0"
pydantic-settings = "^2.6.1"
opencv-python = "^4.10.0.84"
streamlit = "^1.40.1"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
